[feat] dto 및 service, template 수정
